+++ exoskeleton/__init__.py
@@ -4,4 +4,4 @@
 from exoskeleton.__main__ import Exoskeleton
 
 name = "exoskeleton"
-__version__ = "0.6.3"
+__version__ = "0.7.0"